Create hook for tooltip position ref
